requirements.txt
PyPDF2>=3.0.0
openai>=1.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
